Fix bug in sending found names to bus clients under certain circumstances.
If there are 2 client apps looking for the name prefixes of "org.foo" and
"org.foo.bar" when the name "org.foo.goo" comes in, neither client app will be
notified of that name even though "org.foo.goo" matches the prefix "org.foo".
This was the result of an attempt to avoid walking the whole list of name
prefixes.  This change simply walks the whole list of name prefixes.