ALLJOYN-391 Checking that the InterfaceDescription pointer is not NULL after call ing GetInterface
If BusAttachement can not obtain the indicated interface when calling GetInterface it will set the InterfaceDescription pointer to NULL.  The code assumed that the call to GetInterface was always succesful.  This now adds a check to verify that it was successful and returns ER_BUS_NO_SUCH_INTERFACE if unsuccessful.